requirements.txt
# requirements.txt
fastapi==0.78.0
uvicorn==0.18.3
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.21.1
python-multipart==0.0.5
